refactor to datamanager
